+++ tests/containers/bats/podman.pm
@@ -83,17 +83,19 @@ sub run {
     # Patch tests
     run_command "sed -i 's/^PODMAN_RUNTIME=/&$oci_runtime/' test/system/helpers.bash";
     run_command "rm -f contrib/systemd/system/podman-kube@.service.in";
-    # This test fails on systems with GNU tar 1.35 due to
-    # https://bugzilla.suse.com/show_bug.cgi?id=1246607
-    run_command "rm -f test/system/125-import.bats" if (!is_x86_64 && (is_tumbleweed || is_sle('>=16.0')));
-    # This test is flaky on architectures other than x86_64
-    run_command "rm -f test/system/180-blkio.bats" unless is_x86_64;
-    # This test is flaky on ppc64le & s390x
-    run_command "rm -f test/system/220-healthcheck.bats" if (is_ppc64le || is_s390x);
-    # This test is flaky and will fail if system is "full"
-    run_command "rm -f test/system/320-system-df.bats";
-    # This tests needs criu, available only on Tumbleweed
-    run_command "rm -f test/system/520-checkpoint.bats" unless is_tumbleweed;
+    unless (get_var("BATS_TESTS")) {
+        # This test fails on systems with GNU tar 1.35 due to
+        # https://bugzilla.suse.com/show_bug.cgi?id=1246607
+        run_command "rm -f test/system/125-import.bats" if (!is_x86_64 && (is_tumbleweed || is_sle('>=16.0')));
+        # This test is flaky on architectures other than x86_64
+        run_command "rm -f test/system/180-blkio.bats" unless is_x86_64;
+        # This test is flaky on ppc64le & s390x
+        run_command "rm -f test/system/220-healthcheck.bats" if (is_ppc64le || is_s390x);
+        # This test is flaky and will fail if system is "full"
+        run_command "rm -f test/system/320-system-df.bats";
+        # This tests needs criu, available only on Tumbleweed
+        run_command "rm -f test/system/520-checkpoint.bats" unless is_tumbleweed;
+    }
 
     # Compile helpers used by the tests
     run_command "make podman-testing || true", timeout => 600;
